@@ -8,8 +8,8 @@
 import Foundation
 import IndexStoreDB
 
-/// Enumeration of declaration types source kinds utilised when resolving source types.
-public enum SourceKind: String, CaseIterable, Hashable, Equatable {
+/// Enumeration of declaration types source kinds utilized when resolving source types.
+public enum SourceKind: String, CaseIterable, Hashable, Equatable, Sendable {
     /// Represents an unsupported or unknown source kind.
     case unsupported
     /// Represents a module, such as a Swift module or a C++ namespace.
